requirements.txt
Flask_Cors==5.0.0
geopandas==1.0.1
matplotlib==3.9.2
numpy==2.1.2
odc_geo==0.4.8
odc_stac==0.3.10
pandas==2.2.3
planetary_computer==1.0.0
pyproj==3.6.1
pystac==1.11.0
pystac_client==0.8.3
rasterio==1.3.11
Requests==2.32.3
rioxarray==0.17.0
Shapely==2.0.6
xarray==2024.9.0
